% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getShape.R
\name{convertMethFile}
\alias{convertMethFile}
\title{Convert fasta file to methylated file format}
\usage{
convertMethFile(fastaFileName, methPositionFileName)
}
\arguments{
\item{fastaFileName}{The name of the input fasta format file, including
full path to file if it is located outside the current working directory.}

\item{methPositionFileName}{The name of the input position file
indicating the methlation position}
}
\value{
methFileName fasta file containing methylated Cytosine
}
\description{
Convert fasta file to methylated file format
}
\author{
Satyanarayan Rao & Tsu-Pei Chiu
}
