% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\docType{package}
\name{DNAshapeR-package}
\alias{DNAshapeR-package}
\title{DNAshapeR package for high-throughput prediction of DNA shape features}
\description{
The main functions in the package are \code{\link{getFasta}} and
\code{\link{getShape}}. Shape predictions can be additionally plotted with \code{\link{plotShape}}.
See package vignette for examples.
}
\details{
All support questions should be posted to the Bioconductor
support site: \url{support.bioconductor.org}, using DNAshapeR as a tag.
}
\author{
Tsu-Pei Chiu and Federico Comoglio
}
\references{
DNAshapeR reference:

T.-P. Chiu*, F. Comoglio*, T. Zhou, L. Yang, R. Paro, and R. Rohs: DNAshapeR: an R/Bioconductor package for DNA shape prediction and feature encoding (2016). Bioinformatics \url{http://bioinformatics.oxfordjournals.org/content/early/2016/01/09/bioinformatics.btv735} (*equal contributor in alphabetic order)
}
\keyword{package}

