% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/encodeSeqShape.R
\name{encodeKMerSeq}
\alias{encodeKMerSeq}
\title{encode k-mer DNA sequence features}
\usage{
encodeShape(filepath)
}
\arguments{
\item{k}{k-mer sequence}

\item{dnaStringSet}{}
}
\value{
featureVector
}
\description{
encode k-mer DNA sequence features
}
\note{
None
}

