% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/encodeSeqShape.R
\name{encodeNstOrderShape}
\alias{encodeNstOrderShape}
\title{encode n-st order shape features}
\usage{
encodeNstOrderShape
}
\arguments{
\item{n}{n-st order shape}

\item{shapeMatrix}{DESCR}
}
\value{
featureVector
}
\description{
encode n-st order shape features
}
\note{
None
}

