% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/encodeSeqShape.R
\name{encodeSeqShape}
\alias{encodeSeqShape}
\title{encode k-mer DNA sequence and n-st DNA Shape features}
\usage{
encodeSeqShape
}
\arguments{
\item{fastaFileName}{DESCR}

\item{shapeMatrix}{DESCR}

\item{featureNames}{DESCR}
}
\value{
featureVector
}
\description{
encode k-mer DNA sequence and n-st DNA Shape features
}
\note{
None
}
\author{
Tsu-Pei Chiu
}

