% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getFasta.R
\name{getFasta}
\alias{getFasta}
\title{Extract fasta sequence given a set of genomic intervals and a reference genome.}
\usage{
getFasta( GR, BSgenome, width = 1e3, filename = 'tmp.fa' )
}
\arguments{
\item{GR}{}

\item{BSgenome}{}

\item{width}{}

\item{filename}{}
}
\value{
writes a fasta file
}
\description{
Extract fasta sequence given a set of genomic intervals and a reference genome. Intervals are resized to a fixed width.
}
\note{
None
}
\author{
Federico Comoglio
}
\seealso{
\code{\link{}}
}
\keyword{core}

