% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getShape}
\alias{getShape}
\title{Predict DNA shape from FASTA}
\usage{
getShape(filename, shapeType = 'All', parse = TRUE)
}
\arguments{
\item{filename}{DESCR}

\item{shapeType}{DESCR}

\item{parse}{DESCR}
}
\value{
txt file with shape predictions in the working directory
}
\description{
Predict DNA shape from FASTA
}
\note{
None.
}
\author{
Federico Comoglio
}
\keyword{core}

