% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodeSeqShape.R
\name{normalizeShape}
\alias{normalizeShape}
\title{Normalize n-st order shape features}
\usage{
normalizeShape(featureVector, thOrder, shapeType, normalize)
}
\arguments{
\item{featureVector}{A matrix containing encoded features.}

\item{thOrder}{A number indicating n-st order shape encoding}

\item{shapeType}{A character name of shape (MGW, Roll, ProT, HelT) features}

\item{normalize}{A logical indicating whether to perform
normalization. Default to TRUE.}
}
\value{
featureVector A matrix containing encoded features.
}
\description{
Normalize n-st order shape features
}
\author{
Tsu-Pei Chiu
}

