% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parseShape.R
\name{parseShape}
\alias{parseShape}
\title{Parse output files (internal)}
\usage{
parseShape( filename)
}
\arguments{
\item{filename}{DESCR}
}
\value{
Matrix
}
\description{
Parse output files (internal)
}
\note{
None.
}
\author{
Federico Comoglio
}
\keyword{core}

