% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotShape.R
\name{plotShape}
\alias{plotShape}
\title{Plot metaprofiles of DNA shape features}
\usage{
plotShape(shapeMatrix, background = NULL,
colDots = rgb( 0, 0, 1, 0.1),
colDotsBg = rgb( 0, 0, 0, 0.1),
colLine = 'steelblue', colLineBg = 'gray50', cex = 0.5, lwd = 2, ...)
}
\arguments{
\item{shapeMatrix}{A matrix containing DNAshape prediction result}

\item{background}{DESCR}

\item{colDots}{DESCR}

\item{colDotsBg}{DESCR}

\item{colLine}{DESCR}

\item{colLineBg}{DESCR}

\item{cex}{DESCR}

\item{lwd}{DESCR}

\item{...}{DESCR}
}
\value{
Called for its effects
}
\description{
DNAshapeR can be used to generate various graphical representations for
further analysis. The prediction result can be visualized in the form of
plots (Comoglio, et al., 2015), heat maps (Yang, et al., 2014), or
genome browser tracks (Chiu, et al., 2014), or used for the
assembly of feature vectors that include  a user-defined combination of
k-mer and DNA shape features (Zhou, et al., 2015)
}
\examples{

fn <- system.file("extdata", "CGRsample.fa", package = "DNAshapeR")
pred <- getShape(fn)
plotShape(pred$MGW)
plotShape(pred$ProT)
plotShape(pred$Roll)
plotShape(pred$HelT)

}
\author{
Federico Comoglio
}
\keyword{core}

