% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getShape.R
\name{readNonStandardFastaFile}
\alias{readNonStandardFastaFile}
\title{Read the position fasta file}
\usage{
readNonStandardFastaFile(filename)
}
\arguments{
\item{filename}{The name of the input position file
indicating the methlation position}
}
\value{
df dataframe
}
\description{
Read the position fasta file
}
\author{
Satyanarayan Rao & Tsu-Pei Chiu
}
