% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getShape.R
\name{readShape}
\alias{readShape}
\title{Read (parse) DNA shape predictions}
\usage{
readShape(filename)
}
\arguments{
\item{filename}{character name of the file containing shape predictions, including
full path to file if it is located outside the current working directory.}
}
\value{
shapeMatrix matrix containing the shape prediction result
}
\description{
Read DNA shape predictions
}
\examples{
fn <- system.file("extdata", "CGRsample.fa", package = "DNAshapeR")
pred <- readShape(fn)
}
\author{
Federico Comoglio & Tsu-Pei Chiu
}
\keyword{core}

