% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotShape.R
\name{trackShape}
\alias{trackShape}
\title{Plot track view of DNA shape features}
\usage{
trackShape( filename, shapeList )
}
\arguments{
\item{filename}{DESCR}

\item{shapeList}{DESCR}
}
\value{
Called for its effects
}
\description{
Plot track view of DNA shape features
}
\note{
None.
}
\author{
Tsu-Pei Chiu
}
\keyword{core}

