% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodeSeqShape.R
\name{encodeKmerHbond}
\alias{encodeKmerHbond}
\title{encode Hbond}
\usage{
encodeKmerHbond (k, dnaStringSet )
}
\arguments{
\item{k}{k-mer sequence}

\item{dnaStringSet}{dnaStringSet}
}
\value{
featureVector
}
\description{
encode Hbond
}
