% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodeSeqShape.R
\name{encodeNstOrderShape}
\alias{encodeNstOrderShape}
\title{Encode n-st order shape features
DNAshapeR can be used to generate feature vectors for a user-defined model.
The model can be a shape model. There are four structural parameters
including MGW, Roll, ProT and HelT. The second order shape features are
product terms of values for the same category of shape features at adjacent
positions.}
\usage{
encodeNstOrderShape(n, shapeMatrix, shapeType)
}
\arguments{
\item{n}{A number indicating n-st order shape encoding}

\item{shapeMatrix}{A matrix containing DNAshape prediction result}

\item{shapeType}{A character name of shape (MGW, Roll, ProT, HelT) features}
}
\value{
featureVector A matrix containing encoded features. shape feature is
represented as continuous numbers
}
\description{
Encode n-st order shape features
DNAshapeR can be used to generate feature vectors for a user-defined model.
The model can be a shape model. There are four structural parameters
including MGW, Roll, ProT and HelT. The second order shape features are
product terms of values for the same category of shape features at adjacent
positions.
}
\author{
Tsu-Pei Chiu
}

