% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotShape.R
\name{heatShape}
\alias{heatShape}
\title{Plot heatmap of DNA shape features}
\usage{
heatShape(shapeMatrix, nBins, ordRow = NULL, useRaster = TRUE, ... )
}
\arguments{
\item{shapeMatrix}{A matrix containing DNAshape prediction result}

\item{nBins}{DESCR}

\item{ordRow}{DESCR}

\item{useRaster}{DESCR}

\item{...}{DESCR}
}
\value{
Called for its effects
}
\description{
Plot heatmap of DNA shape features
}
\note{
None.
}
\examples{
library(fields)
heatShape(pred$ProT, 20)
heatShape(pred$MGW, 20)
heatShape(pred$Roll[1:500, 1:1980], 20)
heatShape(pred$HelT[1:500, 1:1980], 20)
}
\author{
Federico Comoglio
}
\keyword{core}

