% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotShape.R
\name{heatShape}
\alias{heatShape}
\title{Plot heatmap of DNA shape features}
\usage{
heatShape(shapeMatrix, nBins, ordRow = NULL, useRaster = TRUE, ... )
}
\arguments{
\item{shapeMatrix}{A matrix containing DNAshape prediction result}

\item{nBins}{DESCR}

\item{ordRow}{DESCR}

\item{useRaster}{DESCR}

\item{...}{DESCR}
}
\value{
Called for its effects
}
\description{
Plot heatmap of DNA shape features
}
\examples{

fn <- system.file("extdata", "CGRsample.fa", package = "DNAshapeR")
pred <- getShape(fn)
library(fields)
heatShape(pred$MGW, 20)

}
\author{
Federico Comoglio
}
\keyword{core}

