% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotShape.R
\name{plotShape}
\alias{plotShape}
\title{Plot metaprofiles of DNA shape features}
\usage{
plotShape(shapeMatrix, background = NULL, colDots = rgb( 0, 0, 1, 0.1), colDotsBg = rgb( 0, 0, 0, 0.1), colLine = 'steelblue', colLineBg = 'gray50', cex = 0.5, lwd = 2, ...)
}
\arguments{
\item{shapeMatrix}{DESCR}

\item{background}{DESCR}

\item{colDots}{DESCR}

\item{colDotsBg}{DESCR}

\item{colLine}{DESCR}

\item{colLineBg}{DESCR}

\item{cex}{DESCR}

\item{lwd}{DESCR}

\item{...}{DESCR}
}
\value{
Called for its effects
}
\description{
Plot metaprofiles of DNA shape features
}
\note{
None.
}
\author{
Federico Comoglio
}
\keyword{core}

