% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotShape.R
\name{trackShape}
\alias{trackShape}
\title{Plot track view of DNA shape features}
\usage{
trackShape( filename, shapeList )
}
\arguments{
\item{filename}{The name of the input fasta format file, including
full path to file if it is located outside the current working directory}

\item{shapeList}{A list containing four DNAshape prediction results}
}
\value{
Called for its effects
}
\description{
Plot track view of DNA shape features
}
\note{
None.
}
\examples{
fn2 <- system.file("extdata", "SingleSeqsample.fa", package = "DNAshapeR")
pred2 <- getShape(fn2)
trackShape(fn2, pred2) # Only for single sequence file
}
\author{
Tsu-Pei Chiu
}
\keyword{core}

